+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Input->filter() will call all filters even when the value is `null`, filters are required to be null-safe.
 
 ## [0.1.8] - 2021-09-06
 ### Fixed